+++ src/pages/_app.jsx
@@ -1,6 +1,6 @@
 import Head from 'next/head';
 import { Fragment } from 'react';
-import { AuthProvider } from '../contexts/AuthContext';
+import { AuthProvider } from '../common/contexts/AuthContext';
 import '../styles/styles.scss';
 
 export default function App({ Component, pageProps }) {
