+++ src/pages/posts/index.jsx
@@ -2,10 +2,10 @@ import { useState } from 'react';
 import { Modal } from 'react-responsive-modal';
 import 'react-responsive-modal/styles.css';
 
-import PostApi from '../../api/PostsApi';
-import Table from '../../components/Table/Table';
-import PostEditForm from '../../components/Posts/PostEditForm';
-import PrivateLayout from '../../layouts/PrivateLayout';
+import PostApi from '../../utils/httpClient/api/PostsApi';
+import Table from '../../common/components/Table/Table';
+import PostEditForm from '../../common/components/Posts/PostEditForm';
+import PrivateLayout from '../../common/layouts/PrivateLayout';
 
 const cols = [
     {
@@ -25,18 +25,18 @@ const Posts = ({ posts, errors }) => {
     const openEditModal = (item) => {
         setModalOpen(true);
         setCurrentPost(item);
-    }
+    };
 
     const onCloseModal = () => {
         setModalOpen(false);
         setCurrentPost(null);
-    }
+    };
 
     return (
         <PrivateLayout>
             {errors && <div>Error in fetch</div>}
             {posts && <Table itemsPerPage={15} data={posts} cols={cols} actionHandler={openEditModal} />}
-            {currentPost &&
+            {currentPost && (
                 <Modal open={modalOpen} onClose={onCloseModal} center classNames={{ modal: 'w-75' }}>
                     <h2>Edit post {currentPost.id}</h2>
                     <div className="row pt-3">
@@ -45,7 +45,7 @@ const Posts = ({ posts, errors }) => {
                         </div>
                     </div>
                 </Modal>
-            }
+            )}
         </PrivateLayout>
     );
 };
@@ -56,11 +56,10 @@ export async function getServerSideProps() {
 
     try {
         posts = await PostApi.getPosts();
-
     } catch (e) {
         errors = {
             statusCode: 404
-        }
+        };
     }
 
     return {
