+++ src/pages/users/[id].jsx
@@ -1,10 +1,10 @@
 import { FormProvider, useForm } from 'react-hook-form';
 import { useState } from 'react';
 import { useRouter } from 'next/router';
-import UsersApi from '../../api/UsersApi';
+import UsersApi from '../../utils/httpClient/api/UsersApi';
 
 import styles from './User.module.scss';
-import PrivateLayout from '../../layouts/PrivateLayout';
+import PrivateLayout from '../../common/layouts/PrivateLayout';
 
 const User = ({ user, error }) => {
     const router = useRouter();
@@ -12,7 +12,7 @@ const User = ({ user, error }) => {
 
     const userForm = {
         ...['name', 'username', 'email', 'phone', 'website'].reduce((mem, key) => ({ ...mem, [key]: user[key] }), {})
-    }
+    };
 
     const methods = useForm({
         mode: 'onChange',
@@ -27,9 +27,7 @@ const User = ({ user, error }) => {
         try {
             await UsersApi.saveUser(data);
             router.push('/users');
-        } catch (e) {
-
-        }
+        } catch (e) {}
 
         setSubmitting(false);
     };
@@ -40,10 +38,12 @@ const User = ({ user, error }) => {
 
             <FormProvider {...methods}>
                 <form onSubmit={handleSubmit(onSubmit)}>
-                    {Object.keys(userForm).map(input => {
+                    {Object.keys(userForm).map((input) => {
                         return (
                             <div className="mb-3 row" key={input}>
-                                <label className={`${styles['form-input-name']} col-sm-2 col-form-label`}>{input}</label>
+                                <label className={`${styles['form-input-name']} col-sm-2 col-form-label`}>
+                                    {input}
+                                </label>
                                 <div className="col-sm-10">
                                     <input
                                         type="text"
@@ -55,15 +55,11 @@ const User = ({ user, error }) => {
                                     />
                                 </div>
                             </div>
-                        )
+                        );
                     })}
 
                     <div className="text-end">
-                        <button
-                            disabled={isSubmitting}
-                            type="submit"
-                            className="btn-primary btn w-25 mt-3"
-                        >
+                        <button disabled={isSubmitting} type="submit" className="btn-primary btn w-25 mt-3">
                             {isSubmitting && (
                                 <span
                                     className="mx-2 spinner-border spinner-border-sm"
@@ -77,8 +73,8 @@ const User = ({ user, error }) => {
                 </form>
             </FormProvider>
         </PrivateLayout>
-    )
-}
+    );
+};
 
 export async function getServerSideProps(context) {
     const user = await UsersApi.getUser(context.params.id);
